refactor: Move post composer to Compose folder #332
