+++ src/Core/Fan.WebApp/Manage/Admin/Posts.cshtml
@@ -17,7 +17,7 @@
                         Draft <v-chip v-if="draftCount > 0" small>{{ draftCount }}</v-chip>
                     </v-tab>
                     <v-spacer></v-spacer>
-                    <v-btn small href="/admin/compose" color="pink" class="text-capitalize white--text"
+                    <v-btn small href="/admin/compose/post" color="pink" class="text-capitalize white--text"
                            style="margin:10px 13px 0 0"><v-icon small>add</v-icon> New Post</v-btn>
                 </v-tabs>
             </v-flex>
