+++ src/Core/Fan.WebApp/Manage/Admin/client/README.md
@@ -16,7 +16,6 @@ babel js/admin.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/plugins.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/widgets.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/blog-categories.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
-babel js/blog-compose.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/blog-media.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/blog-settings.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/blog-tags.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
@@ -24,6 +23,7 @@ babel js/setup.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/site-users.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/compose-page.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/compose-pagenav.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
+babel js/compose-post.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/blog-posts.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 babel js/blog-pages.js --out-dir ../../../wwwroot/admin/js --source-maps --watch
 ```
@@ -43,7 +43,6 @@ babel js/admin.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins tr
 babel js/plugins.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 babel js/widgets.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 babel js/blog-categories.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
-babel js/blog-compose.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 babel js/blog-media.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 babel js/blog-settings.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 babel js/blog-tags.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
@@ -51,6 +50,7 @@ babel js/setup.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins tr
 babel js/site-users.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 babel js/compose-page.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 babel js/compose-pagenav.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
+babel js/compose-post.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 babel js/blog-posts.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 babel js/blog-pages.js --out-dir ../../../wwwroot/admin/js --source-maps --plugins transform-remove-console
 ```
