--- src/Core/Fan.WebApp/wwwroot/admin/js/blog-compose.js
@@ -1,2 +1,0 @@
-var app=new Vue({el:"#app",store,mixins:[composeMixin],data:()=>({pubClicked:!1,pubText:"",saveVisible:!1,saveDisabled:!1,saveText:"Save",closing:!0,fieldChanged:!1,drawer:null,panel:[!0,!0,!0],menuDate:!1,mediaDialogVisible:!1,previewUrl:null,postUrl:null,previewDialogVisible:!1,editor:null,snackbar:{show:!1,text:"",color:""},composerUploadProgress:!1}),computed:{disablePubButton(){return 0>=this.post.title.trim().length||this.pubClicked},tok:function(){return document.querySelector("input[name=\"__RequestVerificationToken\"][type=\"hidden\"]").value},headers(){return{headers:{"XSRF-TOKEN":this.tok}}},payload:function(){return{id:this.post.id,postDate:this.post.postDate,categoryId:this.post.categoryId,tags:this.post.tags,slug:this.post.slug,excerpt:this.post.excerpt,title:this.post.title,body:this.editor.getData()}},selectedImages(){return this.$store.state.selectedImages}},mounted(){this.pubText=this.post.published?"Update":"Publish",this.initWindowDnd(),this.initEditor()},methods:{initWindowDnd(){window.addEventListener("dragenter",function(){document.querySelector("#dropzone").style.visibility="",document.querySelector("#dropzone").style.opacity=1}),window.addEventListener("dragleave",function(a){a.preventDefault(),document.querySelector("#dropzone").style.visibility="hidden",document.querySelector("#dropzone").style.opacity=0}),window.addEventListener("dragover",function(a){a.preventDefault(),document.querySelector("#dropzone").style.visibility="",document.querySelector("#dropzone").style.opacity=1});let a=this;window.addEventListener("drop",function(b){b.preventDefault(),document.querySelector("#dropzone").style.visibility="hidden",document.querySelector("#dropzone").style.opacity=0,a.mediaDialogVisible||a.dragFilesUpload(b.dataTransfer.files)})},dragFilesUpload(a){if(!a.length)return;this.composerUploadProgress=!0;const b=new FormData;Array.from(Array(a.length).keys()).map(c=>b.append("images",a[c])),this.sendImages(b)},sendImages(a){axios.post("/admin/media?handler=image",a,this.$root.headers).then(a=>{0<a.data.images.length?(this.insertImagesToEditor(a.data.images),this.$root.toast("Image uploaded.")):this.$root.toastError("Some files were not uploaded due to file type or size not supported."),this.composerUploadProgress=!1,this.$store.dispatch("emptyErrMsg")}).catch(()=>{this.composerUploadProgress=!1,this.$root.toastError("Image upload failed.")})},initEditor(){let a,b=this;ClassicEditor.create(document.querySelector("#editor"),{autosave:{save(c){clearTimeout(a),b.post.published||(b.saveVisible=!0,b.saveDisabled=!1,b.saveText="Save",a=setTimeout(b.saveDraft,b.autosaveInterval))}}}).then(a=>{b.editor=a}).catch(()=>{})},onFieldsChange(){this.fieldChanged=!0;this.post.published||(this.saveVisible=!0,this.saveDisabled=!1,this.saveText="Save")},saveDraft(){this.saveVisible=!0,this.saveDisabled=!0,this.saveText="Saving...",axios.post("/admin/compose?handler=save",this.payload,{headers:{"XSRF-TOKEN":this.tok}}).then(a=>{this.post.id=a.data.id,this.post.slug=a.data.slug,this.post.draftDate=a.data.draftDate,this.post.isDraft=!0,window.location.href.endsWith("/compose")&&history.replaceState({},null,window.location.href+`/${this.post.id}`)}).catch(()=>{}),this.fieldChanged=!1,this.saveText="Saved"},preview(){this.previewDialogVisible=!0,this.payload.body=this.editor.getData(),axios.post("/admin/compose?handler=preview",this.payload,{headers:{"XSRF-TOKEN":this.tok}}).then(a=>{this.previewUrl=a.data,this.postUrl=this.previewUrl.replace("preview/","")}).catch(()=>{})},closePreview(){this.previewDialogVisible=!1,this.previewUrl=null,this.postUrl=null},revert(){this.post.published=!1,this.pubText=this.post.published?"Update":"Publish",this.saveDraft()},pub(){this.closing=!1,this.pubClicked=!0,this.pubText=this.post.published?"Updating...":"Publishing...";const a=this.post.published?"/admin/compose?handler=update":"/admin/compose?handler=publish";axios.post(a,this.payload,{headers:{"XSRF-TOKEN":this.tok}}).then(a=>{window.location.replace(a.data)}).catch(()=>{})},insertImages(){this.insertImagesToEditor(this.selectedImages),this.mediaDialogVisible=!1,this.$store.dispatch("emptySelectedImages"),this.$store.dispatch("emptyErrMsg")},insertImagesToEditor(a){let b="";a.forEach(a=>{b+=`<figure class="image"><img src="${a.urlMedium}" alt="${a.alt}" title="${a.title}"><figcaption>${a.caption}</figcaption></figure>`,a.selected=!1});const c=this.editor.data.processor.toView(b),d=this.editor.data.toModel(c);this.editor.model.insertContent(d,this.editor.model.document.selection)},closeMediaDialog(){this.mediaDialogVisible=!1,this.selectedImages.forEach(a=>a.selected=!1),this.$store.dispatch("emptySelectedImages"),this.$store.dispatch("emptyErrMsg")},titleEnter(){this.post.title=this.post.title.replace(/\n/g," ")},toast(a,b="silver"){this.snackbar.show=!0,this.snackbar.text=a,this.snackbar.color=b},toastError(a){this.snackbar.show=!0,this.snackbar.text=a,this.snackbar.color="red"}}});
-//# sourceMappingURL=blog-compose.js.map
