+++ src/Themes/Clarity/Views/Shared/Header.cshtml
@@ -14,9 +14,15 @@
             <div class="dropdown">
                 <a class="p-2 text-dark nav-link" href="#" id="adminDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Admin</a>
                 <div class="dropdown-menu dropdown-menu-right" aria-labelledby="adminDropdown">
-                    <a class="dropdown-item" href="/admin/posts">Posts</a>
+                    <div class="d-flex justify-content-between">
+                        <a class="dropdown-item" href="/admin/posts">Posts</a>
+                        <a class="dropdown-item text-right pr-3" href="/admin/compose/post"><i class="icon-pencil-square-o" aria-hidden="true"></i></a>
+                    </div>
                     <div class="dropdown-divider m-0"></div>
-                    <a class="dropdown-item" href="/admin/pages">Pages</a>
+                    <div class="d-flex justify-content-between">
+                        <a class="dropdown-item" href="/admin/pages">Pages</a>
+                        <a class="dropdown-item text-right pr-3" href="/admin/compose/page"><i class="icon-pencil-square-o" aria-hidden="true"></i></a>
+                    </div>
                     <div class="dropdown-divider m-0"></div>
                     <a class="dropdown-item" href="/admin/media">Media</a>
                     <div class="dropdown-divider m-0"></div>
@@ -37,16 +43,11 @@
                     <a class="dropdown-item" href="/admin/settings">Settings</a>
                 </div>
             </div>
-
             <div class="dropdown">
                 <a class="nav-link" href="#" id="avatarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                     <img gravatar-email="@Model.CurrentUser.Email" title="@Model.CurrentUser.DisplayName" class="avatar" />
                 </a>
                 <div class="dropdown-menu dropdown-menu-right" aria-labelledby="avatarDropdown">
-                    <a class="dropdown-item" href="/admin/compose"><i class="icon-pencil-square-o" aria-hidden="true"></i> Post</a>
-                    <div class="dropdown-divider m-0"></div>
-                    <a class="dropdown-item" href="/admin/compose/page"><i class="icon-pencil-square-o" aria-hidden="true"></i> Page</a>
-                    <div class="dropdown-divider m-0"></div>
                     <form asp-area="" asp-controller="Home" asp-action="Logout" method="post" id="logoutForm">
                         <button type="submit" class="dropdown-item btn btn-link navbar-btn navbar-link"><i class="icon-power-off" aria-hidden="true"></i> Logout</button>
                     </form>
